+++ async_firebase/client.py
@@ -7,18 +7,15 @@ to authorize request which is being made to Firebase.
 import json
 import logging
 import typing as t
-import uuid
 from dataclasses import replace
 from datetime import datetime, timedelta
 from email.mime.multipart import MIMEMultipart
 from email.mime.nonmultipart import MIMENonMultipart
-from pathlib import PurePath
-from urllib.parse import urlencode, urljoin
+from urllib.parse import urljoin
 
 import httpx
-from google.oauth2 import service_account  # type: ignore
 
-import pkg_resources  # type: ignore
+from async_firebase.base import AsyncClientBase
 from async_firebase.encoders import aps_encoder
 from async_firebase.messages import (
     AndroidConfig,
@@ -27,10 +24,10 @@ from async_firebase.messages import (
     APNSPayload,
     Aps,
     ApsAlert,
-    FcmPushMulticastResponse,
-    FcmPushResponse,
+    FCMBatchResponse,
+    FCMResponse,
     Message,
-    Notification,
+    MulticastMessage,
     PushNotification,
     WebpushConfig,
     WebpushFCMOptions,
@@ -38,53 +35,22 @@ from async_firebase.messages import (
     WebpushNotificationAction,
 )
 from async_firebase.utils import (
-    FcmPushMulticastResponseHandler,
-    FcmPushResponseHandler,
+    FCMBatchResponseHandler,
+    FCMResponseHandler,
     cleanup_firebase_message,
     serialize_mime_message,
 )
 
 DEFAULT_TTL = 604800
-MULTICAST_MESSAGE_MAX_DEVICE_TOKENS = 500
+BATCH_MAX_MESSAGES = MULTICAST_MESSAGE_MAX_DEVICE_TOKENS = 500
 
 
-class AsyncFirebaseClient:
+class AsyncFirebaseClient(AsyncClientBase):
     """Async wrapper for Firebase Cloud Messaging.
 
     The AsyncFirebaseClient relies on Service Account to enable us making a request. To get more about Service Account
     please refer to https://firebase.google.com/support/guides/service-accounts
     """
-
-    BASE_URL: str = "https://fcm.googleapis.com"
-    TOKEN_URL: str = "https://oauth2.googleapis.com/token"
-    FCM_ENDPOINT: str = "/v1/projects/{project_id}/messages:send"
-    FCM_BATCH_ENDPOINT: str = "/batch"
-    # A list of accessible OAuth 2.0 scopes can be found https://developers.google.com/identity/protocols/oauth2/scopes.
-    SCOPES: t.List[str] = [
-        "https://www.googleapis.com/auth/cloud-platform",
-    ]
-
-    def __init__(
-        self,
-        credentials: t.Optional[service_account.Credentials] = None,
-        scopes: t.Optional[t.List[str]] = None,
-    ) -> None:
-        """
-        :param credentials: instance of ``google.oauth2.service_account.Credentials``.
-            Usually, you'll create these credentials with one of the helper constructors. To create credentials using a
-            Google service account private key JSON file::
-
-                self.creds_from_service_account_file('service-account.json')
-
-            Or if you already have the service account file loaded::
-
-                service_account_info = json.load(open('service_account.json'))
-                self.creds_from_service_account_info(service_account_info)
-
-        :param scopes: user-defined scopes to request during the authorization grant.
-        """
-        self._credentials: service_account.Credentials = credentials
-        self.scopes: t.List[str] = scopes or self.SCOPES
 
     @staticmethod
     def assemble_push_notification(
@@ -114,51 +80,6 @@ class AsyncFirebaseClient:
             logging.warning("No data has been provided to construct push notification payload")
             raise ValueError("``messages.PushNotification`` cannot be assembled as data has not been provided")
         return push_notification
-
-    def creds_from_service_account_info(self, service_account_info: t.Dict[str, str]) -> None:
-        """
-        Creates a Credentials instance from parsed service account info.
-
-        :param service_account_info: the service account info in Google format.
-        """
-        self._credentials = service_account.Credentials.from_service_account_info(
-            info=service_account_info, scopes=self.scopes
-        )
-
-    def creds_from_service_account_file(self, service_account_filename: t.Union[str, PurePath]) -> None:
-        """
-        Creates a Credentials instance from a service account json file.
-
-        :param service_account_filename: the path to the service account json file.
-        """
-        if isinstance(service_account_filename, PurePath):
-            service_account_filename = str(service_account_filename)
-
-        logging.debug("Creating credentials from file: %s", service_account_filename)
-        self._credentials = service_account.Credentials.from_service_account_file(
-            filename=service_account_filename, scopes=self.scopes
-        )
-
-    async def _get_access_token(self) -> str:
-        """Get OAuth 2 access token."""
-        if self._credentials.valid:
-            return self._credentials.token
-
-        headers = {"Content-Type": "application/x-www-form-urlencoded"}
-        data = urlencode(
-            {
-                "grant_type": "urn:ietf:params:oauth:grant-type:jwt-bearer",
-                "assertion": self._credentials._make_authorization_grant_assertion(),
-            }
-        ).encode("utf-8")
-
-        async with httpx.AsyncClient() as client:
-            response: httpx.Response = await client.post(self.TOKEN_URL, data=data, headers=headers)
-            response_data = response.json()
-
-        self._credentials.expiry = datetime.utcnow() + timedelta(seconds=response_data["expires_in"])
-        self._credentials.token = response_data["access_token"]
-        return self._credentials.token
 
     @staticmethod
     def build_android_config(  # pylint: disable=too-many-locals
@@ -411,198 +332,144 @@ class AsyncFirebaseClient:
             fcm_options=WebpushFCMOptions(link=link) if link else None,
         )
 
-    @staticmethod
-    def _get_request_id():
-        """Generate unique request ID."""
-        return str(uuid.uuid4())
-
-    @staticmethod
-    def _serialize_batch_request(request: httpx.Request) -> str:
-        """
-        Convert an HttpRequest object into a string.
-
-        :param request: `httpx.Request`, the request to serialize.
-        :return: a string in application/http format.
-        """
-        status_line = f"{request.method} {request.url.path} HTTP/1.1\n"
-        major, minor = request.headers.get("content-type", "application/json").split("/")
-        msg = MIMENonMultipart(major, minor)
-        headers = request.headers.copy()
-
-        # MIMENonMultipart adds its own Content-Type header.
-        if "content-type" in headers:
-            del headers["content-type"]
-
-        for key, value in headers.items():
-            msg[key] = value
-        msg.set_unixfrom(None)  # type: ignore
-
-        if request.content is not None:
-            msg.set_payload(request.content)
-            msg["content-length"] = str(len(request.content))
-
-        body = serialize_mime_message(msg, max_header_len=0)
-        return f"{status_line}{body}"
-
-    async def _prepare_headers(self):
-        """Prepare HTTP headers that will be used to request Firebase Cloud Messaging."""
-        logging.debug("Preparing HTTP headers for all the subsequent requests")
-        access_token: str = await self._get_access_token()
-        return {
-            "Authorization": f"Bearer {access_token}",
-            "Content-Type": "application/json; UTF-8",
-            "X-Request-Id": self._get_request_id(),
-            "X-GOOG-API-FORMAT-VERSION": "2",
-            "X-FIREBASE-CLIENT": "async-firebase/{0}".format(pkg_resources.get_distribution("async-firebase").version),
-        }
-
-    async def push(  # pylint: disable=too-many-locals
-        self,
-        device_token: str,
-        *,
-        android: t.Optional[AndroidConfig] = None,
-        apns: t.Optional[APNSConfig] = None,
-        condition: t.Optional[str] = None,
-        data: t.Optional[t.Dict[str, str]] = None,
-        notification: t.Optional[Notification] = None,
-        topic: t.Optional[str] = None,
-        webpush: t.Optional[WebpushConfig] = None,
-        dry_run: bool = False,
-    ) -> FcmPushResponse:
+    async def send(self, message: Message, *, dry_run: bool = False) -> FCMResponse:
         """
         Send push notification.
 
-        :param device_token: device token allows to send targeted notifications to a particular instance of app.
-        :param android: as instance of ``messages.AndroidConfig`` that contains Android-specific options.
-        :param apns: as instance of ``messages.APNSConfig`` that contains iOS-specific options.
-        :param condition: the Firebase condition to which the message should be sent.
-        :param data: a dictionary of data fields. All keys and values in the dictionary must be strings.
-        :param notification: an instance of ``messages.Notification`` that contains a notification that can be included
-            in a resulting message.
-        :param topic: name of the Firebase topic to which the message should be sent.
-        :param webpush: an instance of ``messages.WebpushConfig``.
+        :param message: the message that has to be sent.
         :param dry_run: indicating whether to run the operation in dry run mode (optional). Flag for testing the request
             without actually delivering the message. Default to ``False``.
 
         :raises:
 
-            ValueError is ``messages.PushNotification`` payload cannot be assembled
+            ValueError if ``messages.PushNotification`` payload cannot be assembled
 
-        :return: response from Firebase. Example of response:
+        :return: instance of ``messages.FCMResponse``
 
-            success::
+            Example of raw response:
 
-                {
-                    'name': 'projects/mobile-app/messages/0:1612788010922733%7606eb247606eb24'
-                }
+                success::
 
-            failure::
-
-                {
-                    'error': {
-                        'code': 400,
-                        'details': [
-                            {
-                                '@type': 'type.googleapis.com/google.rpc.BadRequest',
-                                'fieldViolations': [
-                                    {
-                                        'description': 'Value type for APS key [badge] is a number.',
-                                        'field': 'message.apns.payload.aps.badge'
-                                    }
-                                ]
-                            },
-                            {
-                                '@type': 'type.googleapis.com/google.firebase.fcm.v1.FcmError',
-                                'errorCode': 'INVALID_ARGUMENT'
-                            }
-                        ],
-                        'message': 'Value type for APS key [badge] is a number.',
-                        'status': 'INVALID_ARGUMENT'
+                    {
+                        'name': 'projects/mobile-app/messages/0:1612788010922733%7606eb247606eb24'
                     }
-                }
+
+                failure::
+
+                    {
+                        'error': {
+                            'code': 400,
+                            'details': [
+                                {
+                                    '@type': 'type.googleapis.com/google.rpc.BadRequest',
+                                    'fieldViolations': [
+                                        {
+                                            'description': 'Value type for APS key [badge] is a number.',
+                                            'field': 'message.apns.payload.aps.badge'
+                                        }
+                                    ]
+                                },
+                                {
+                                    '@type': 'type.googleapis.com/google.firebase.fcm.v1.FcmError',
+                                    'errorCode': 'INVALID_ARGUMENT'
+                                }
+                            ],
+                            'message': 'Value type for APS key [badge] is a number.',
+                            'status': 'INVALID_ARGUMENT'
+                        }
+                    }
         """
-        message = Message(
-            token=device_token,
-            data=data or {},
-            notification=notification,
-            android=android,
-            webpush=webpush,
-            apns=apns,
-            topic=topic,
-            condition=condition,
-        )
+        push_notification = self.assemble_push_notification(apns_config=message.apns, dry_run=dry_run, message=message)
 
-        push_notification = self.assemble_push_notification(apns_config=apns, dry_run=dry_run, message=message)
-
-        response = await self._send_request(
+        response = await self.send_request(
             uri=self.FCM_ENDPOINT.format(project_id=self._credentials.project_id),  # type: ignore
             json_payload=push_notification,
-            headers=await self._prepare_headers(),
-            response_handler=FcmPushResponseHandler(),
+            response_handler=FCMResponseHandler(),
         )
-        if not isinstance(response, FcmPushResponse):
+        if not isinstance(response, FCMResponse):
             raise ValueError("Wrong return type, perhaps because of a response handler misuse.")
         return response
 
-    async def push_multicast(
+    async def send_multicast(
         self,
-        device_tokens: t.Union[t.List[str], t.Tuple[str]],
+        multicast_message: MulticastMessage,
         *,
-        android: t.Optional[AndroidConfig] = None,
-        apns: t.Optional[APNSConfig] = None,
-        data: t.Optional[t.Dict[str, str]] = None,
-        notification: t.Optional[Notification] = None,
-        webpush: t.Optional[WebpushConfig] = None,
         dry_run: bool = False,
-    ) -> FcmPushMulticastResponse:
+    ) -> FCMBatchResponse:
         """
         Send Multicast push notification.
 
-        :param device_tokens: the list of device tokens to send targeted notifications to a set of instances of app.
+        :param multicast_message: multicast message to send targeted notifications to a set of instances of app.
             May contain up to 500 device tokens.
-        :param android: as instance of ``messages.AndroidConfig`` that contains Android-specific options.
-        :param apns: as instance of ``messages.APNSConfig`` that contains iOS-specific options.
-        :param data: a dictionary of data fields. All keys and values in the dictionary must be strings.
-        :param notification: an instance of ``messages.Notification`` that contains a notification that can be included
-            in a resulting message.
-        :param webpush: an instance of ``messages.WebpushConfig``.
         :param dry_run: indicating whether to run the operation in dry run mode (optional). Flag for testing the request
             without actually delivering the message. Default to ``False``.
+
+        :raises:
+
+            ValueError if ``messages.PushNotification`` payload cannot be assembled
+            ValueError if ``messages.MulticastMessage`` contains more than MULTICAST_MESSAGE_MAX_DEVICE_TOKENS
+
+        :return: instance of ``messages.FCMBatchResponse``
         """
 
-        if len(device_tokens) > MULTICAST_MESSAGE_MAX_DEVICE_TOKENS:
+        if len(multicast_message.tokens) > MULTICAST_MESSAGE_MAX_DEVICE_TOKENS:
             raise ValueError(
                 f"A single ``messages.MulticastMessage`` may contain up to {MULTICAST_MESSAGE_MAX_DEVICE_TOKENS} "
                 "device tokens."
             )
 
+        messages = [
+            Message(
+                token=token,
+                data=multicast_message.data,
+                notification=multicast_message.notification,
+                android=multicast_message.android,
+                webpush=multicast_message.webpush,
+                apns=multicast_message.apns,
+            )
+            for token in multicast_message.tokens
+        ]
+
+        return await self.send_all(messages, dry_run=dry_run)
+
+    async def send_all(
+        self,
+        messages: t.Union[t.List[Message], t.Tuple[Message]],
+        *,
+        dry_run: bool = False,
+    ) -> FCMBatchResponse:
+        """
+        Send the given messages to FCM in a single batch.
+
+        :param messages: the list of messages to send.
+        :param dry_run: indicating whether to run the operation in dry run mode (optional). Flag for testing the request
+            without actually delivering the message. Default to ``False``.
+        :returns: instance of ``messages.FCMBatchResponse``
+        """
+
+        if len(messages) > BATCH_MAX_MESSAGES:
+            raise ValueError(f"A list of messages must not contain more than {BATCH_MAX_MESSAGES} elements")
+
         multipart_message = MIMEMultipart("mixed")
         # Message should not write out it's own headers.
         setattr(multipart_message, "_write_headers", lambda self: None)
 
-        for device_token in device_tokens:
+        for message in messages:
             msg = MIMENonMultipart("application", "http")
             msg["Content-Transfer-Encoding"] = "binary"
-            msg["Content-ID"] = self._get_request_id()
+            msg["Content-ID"] = self.get_request_id()
             push_notification = self.assemble_push_notification(
-                apns_config=apns,
+                apns_config=message.apns,
                 dry_run=dry_run,
-                message=Message(
-                    token=device_token,
-                    data=data or {},
-                    notification=notification,
-                    android=android,
-                    webpush=webpush,
-                    apns=apns,
-                ),
+                message=message,
             )
-            body = self._serialize_batch_request(
+            body = self.serialize_batch_request(
                 httpx.Request(
                     method="POST",
                     url=urljoin(
                         self.BASE_URL, self.FCM_ENDPOINT.format(project_id=self._credentials.project_id)  # type: ignore
                     ),
-                    headers=await self._prepare_headers(),
+                    headers=await self.prepare_headers(),
                     content=json.dumps(push_notification),
                 )
             )
@@ -612,57 +479,12 @@ class AsyncFirebaseClient:
         # encode the body: note that we can't use `as_string`, because it plays games with `From ` lines.
         body = serialize_mime_message(multipart_message, mangle_from=False)
 
-        batch_response = await self._send_request(
+        batch_response = await self.send_request(
             uri=self.FCM_BATCH_ENDPOINT,
             content=body,
             headers={"Content-Type": f"multipart/mixed; boundary={multipart_message.get_boundary()}"},
-            response_handler=FcmPushMulticastResponseHandler(),
+            response_handler=FCMBatchResponseHandler(),
         )
-        if not isinstance(batch_response, FcmPushMulticastResponse):
+        if not isinstance(batch_response, FCMBatchResponse):
             raise ValueError("Wrong return type, perhaps because of a response handler misuse.")
         return batch_response
-
-    async def _send_request(
-        self,
-        uri: str,
-        response_handler: t.Union[FcmPushResponseHandler, FcmPushMulticastResponseHandler],
-        json_payload: t.Optional[t.Dict[str, t.Any]] = None,
-        headers: t.Optional[t.Dict[str, str]] = None,
-        content: t.Union[str, bytes, t.Iterable[bytes], t.AsyncIterable[bytes], None] = None,
-    ) -> t.Union[FcmPushResponse, FcmPushMulticastResponse]:
-        """
-        Sends an HTTP call using the ``httpx`` library.
-
-        :param uri: URI to be requested.
-        :param json_payload: request JSON payload
-        :param headers: request headers.
-        :param content: request content
-        :return: HTTP response
-        """
-        async with httpx.AsyncClient(base_url=self.BASE_URL) as client:
-            logging.debug(
-                "Requesting POST %s, payload: %s, content: %s, headers: %s",
-                urljoin(self.BASE_URL, self.FCM_ENDPOINT.format(project_id=self._credentials.project_id)),
-                json_payload,
-                content,
-                headers,
-            )
-            try:
-                raw_fcm_response: httpx.Response = await client.post(
-                    uri,
-                    json=json_payload,
-                    headers=headers or await self._prepare_headers(),
-                    content=content,
-                )
-                raw_fcm_response.raise_for_status()
-            except httpx.HTTPError as exc:
-                response = response_handler.handle_error(exc)
-            else:
-                logging.debug(
-                    "Response Code: %s, Time spent to make a request: %s",
-                    raw_fcm_response.status_code,
-                    raw_fcm_response.elapsed,
-                )
-                response = response_handler.handle_response(raw_fcm_response)
-
-        return response