+++ async_firebase/messages.py
@@ -299,7 +299,7 @@ class Message:
     condition: the Firebase condition to which the message should be sent (optional).
     """
 
-    token: t.Optional[str]
+    token: str
     data: t.Dict[str, str] = field(default_factory=dict)
     notification: t.Optional[Notification] = field(default=None)
     android: t.Optional[AndroidConfig] = field(default=None)
@@ -307,6 +307,28 @@ class Message:
     apns: t.Optional[APNSConfig] = field(default=None)
     topic: t.Optional[str] = None
     condition: t.Optional[str] = None
+
+
+@dataclass
+class MulticastMessage:
+    """
+    A message that can be sent to multiple tokens via Firebase.
+
+    Attributes:
+    tokens: a list of registration tokens of targeted devices.
+    data: a dictionary of data fields (optional). All keys and values in the dictionary must be strings.
+    notification: an instance of ``messages.Notification`` (optional).
+    android: an instance of ``messages.AndroidConfig`` (optional).
+    webpush: an instance of ``messages.WebpushConfig`` (optional).
+    apns: an instance of ``messages.ApnsConfig`` (optional).
+    """
+
+    tokens: t.List[str]
+    data: t.Dict[str, str] = field(default_factory=dict)
+    notification: t.Optional[Notification] = field(default=None)
+    android: t.Optional[AndroidConfig] = field(default=None)
+    webpush: t.Optional[WebpushConfig] = field(default=None)
+    apns: t.Optional[APNSConfig] = field(default=None)
 
 
 @dataclass
@@ -322,7 +344,7 @@ class PushNotification:
     validate_only: t.Optional[bool] = field(default=False)
 
 
-class FcmPushResponse:
+class FCMResponse:
     """The response received from an individual batched request to the FCM API.
 
     The interface of this object is compatible with SendResponse object of
@@ -332,7 +354,7 @@ class FcmPushResponse:
     def __init__(
         self, fcm_response: t.Optional[t.Dict[str, str]] = None, exception: t.Optional[AsyncFirebaseError] = None
     ):
-        """Inits FcmPushResponse object.
+        """Inits FCMResponse object.
 
         :param fcm_response: a dictionary with the data that FCM returns as a payload
         :param exception: an exception that may happen when communicating with FCM
@@ -346,24 +368,24 @@ class FcmPushResponse:
         return self.message_id is not None and not self.exception
 
 
-class FcmPushMulticastResponse:
+class FCMBatchResponse:
     """The response received from a batch request to the FCM API.
 
     The interface of this object is compatible with BatchResponse object of
     the Google's firebase-admin-python package.
     """
 
-    def __init__(self, responses: t.List[FcmPushResponse]):
-        """Inits FcmPushMulticastResponse.
+    def __init__(self, responses: t.List[FCMResponse]):
+        """Inits FCMBatchResponse.
 
-        :param responses: a list of FcmPushResponse objects
+        :param responses: a list of FCMResponse objects
         """
         self._responses = responses
         self._success_count = len([resp for resp in responses if resp.success])
 
     @property
     def responses(self):
-        """A list of ``FcmPushResponse`` objects (possibly empty)."""
+        """A list of ``FCMResponse`` objects (possibly empty)."""
         return self._responses
 
     @property